+++ backend/src/utils/cloudinary.ts
@@ -1,14 +1,13 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-const uploadOnCloudinary = async (localFilePath :string) => {
+const uploadOnCloudinary = async (localFilePath: string) => {
   try {
     if (!localFilePath) return null;
 
@@ -17,13 +16,13 @@ const uploadOnCloudinary = async (localFilePath :string) => {
     });
     //uploading success
     console.log("uploaded to cloudinary successfully", response.url);
-    return response
+    return response;
   } catch (err) {
-
     console.error("Cloudinary upload error:", err);
-    fs.unlinkSync(localFilePath)
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
   }
 };
 
-
-export default  uploadOnCloudinary ;
+export default uploadOnCloudinary;
